% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpSolve_Class.R
\docType{class}
\name{lpSolve-class}
\alias{lpSolve-class}
\title{\code{lpSolve} Class Definition}
\description{
\code{lpSolve} Class Definition
}
\section{Slots}{

\describe{
\item{\code{modelname}}{name for model, type character (optional)}

\item{\code{modelsense}}{model sense. legal values are \code{min, max} (optional: default max)}

\item{\code{A}}{model constraints. numeric matrix}

\item{\code{obj}}{objective function. type numeric. length is 1 or ncols of constraints}

\item{\code{ub}}{upper bound. type numeric. length is 1 or ncols of constraints (optional:
default value is Inf)}

\item{\code{lb}}{lower bound. type numeric. length is 1 or ncols of constraints (optional:
default valueis 0)}

\item{\code{rhs}}{RHS constraints. type numeric. length is 1 or nrowss of constraints}

\item{\code{sense}}{sense of constraints. legal values are \code{"<=", "=", ">=", "free"}
length is 1 or nrowss of constraints}
}}

