% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpSolve_Solve_lpSolveAPI.R
\name{lpSolveSolve}
\alias{lpSolveSolve}
\title{Solve method for lpSolve Object}
\usage{
lpSolveSolve(a)
}
\arguments{
\item{a}{lpSolve object to be solved}
}
\description{
Using the lpSolveAPI package solves a lpSolve Object
}

