% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpSolve_Class.R
\name{lpSolveSummary}
\alias{lpSolveSummary}
\alias{summary}
\title{Summary}
\usage{
lpSolveSummary(object)
}
\arguments{
\item{object}{S4 Object to display}
}
\description{
Summary method lpSolve Object
}

